requirements.txt
Django==4.2
psycopg2-binary==2.9.6
Pillow==10.0.0
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0